Refactor/hexagonal architecture

+++ backend/school-management/app/Core/Application/UseCases/Payments/Student/Cards/DeletePaymentMethoUseCase.php
@@ -3,8 +3,10 @@
 namespace App\Core\Application\UseCases\Payments\Student\Cards;
 
 use App\Core\Domain\Repositories\Command\Payments\PaymentMethodRepInterface;
-use App\Core\Domain\Repositories\Command\Stripe\StripeGatewayInterface;
 use App\Core\Domain\Repositories\Query\Payments\PaymentMethodQueryRepInterface;
+use App\Core\Domain\Repositories\Stripe\StripeGatewayInterface;
+use App\Exceptions\NotFound\PaymentMethodNotFoundException;
+use Illuminate\Support\Facades\DB;
 
 class DeletePaymentMethoUseCase
 {
@@ -16,11 +18,20 @@ class DeletePaymentMethoUseCase
     {
     }
 
-    public function execute(string $paymentMethodId): bool
+    public function execute(int $paymentMethodId): bool
     {
         $paymentMethod=$this->pmqRepo->findById($paymentMethodId);
-        $this->stripe->deletePaymentMethod($paymentMethod->stripe_payment_method_id);
-        $this->pmRepo->delete($paymentMethod->id);
+        if (! $paymentMethod) {
+            throw new PaymentMethodNotFoundException();
+        }
+        DB::transaction(function () use ($paymentMethod) {
+            $this->stripe->deletePaymentMethod(
+                $paymentMethod->stripe_payment_method_id
+            );
+
+            $this->pmRepo->delete($paymentMethod->id);
+        });
+
         return true;
     }
 }
